+++ shuup/front/urls.py
@@ -22,7 +22,10 @@ from .views.category import CategoryView
 from .views.checkout import get_checkout_view
 from .views.dashboard import DashboardView
 from .views.index import IndexView
-from .views.misc import toggle_all_seeing
+from .views.misc import (
+    force_anonymous_contact, force_company_contact, force_person_contact,
+    toggle_all_seeing
+)
 from .views.order import OrderCompleteView
 from .views.payment import ProcessPaymentView
 from .views.product import ProductDetailView
@@ -55,6 +58,9 @@ urlpatterns = [
     url(r'^basket/$', csrf_exempt(BasketView.as_view()), name='basket'),
     url(r'^dashboard/$', login_required(DashboardView.as_view()), name="dashboard"),
     url(r'^toggle-allseeing/$', login_required(toggle_all_seeing), name="toggle-all-seeing"),
+    url(r'^force-anonymous-contact/$', login_required(force_anonymous_contact), name="force-anonymous-contact"),
+    url(r'^force-company-contact/$', login_required(force_company_contact), name="force-company-contact"),
+    url(r'^force-person-contact/$', login_required(force_person_contact), name="force-person-contact"),
     url(r'^order/payment/(?P<pk>.+?)/(?P<key>.+?)/$',
         csrf_exempt(ProcessPaymentView.as_view()),
         kwargs={"mode": "payment"},